Moved CreateModelOptions from separate header into UnityPrepareRenderResources.h
