+++ native~/Runtime/src/UnityPrepareRendererResources.h
@@ -4,10 +4,24 @@
 
 #include <Cesium3DTilesSelection/IPrepareRendererResources.h>
 
+#include <DotNet/CesiumForUnity/Cesium3DTileset.h>
 #include <DotNet/UnityEngine/GameObject.h>
 
 namespace CesiumForUnityNative {
 
+struct CreateModelOptions {
+  /**
+   * Whether to ignore the KHR_materials_unlit extension in the model. If this
+   * is true and the extension is present, then flat normals will be generated
+   * for the model as it loads.
+   */
+  bool ignoreKHRMaterialUnlit = false;
+
+  CreateModelOptions() = default;
+  CreateModelOptions(
+      const DotNet::CesiumForUnity::Cesium3DTileset& tilesetComponent)
+      : ignoreKHRMaterialUnlit(tilesetComponent.ignoreKHRMaterialsUnlit()) {}
+};
 /**
  * @brief Information about how a given glTF primitive was converted into
  * Unity MeshData.
